+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v0.3.0
 ### Added
-- Dungeon Quest dropdown filter
+- Dungeon Quest dropdown filters: Faction and Dungeon
+- Dungeon Quest filter now hides options without quests for the selected faction
 ### Changed
 - Dungeon Quest effort now Low, Medium and High
 - Dungeon Quest now shows Dungeon Level range